Fix websocket reconnection when a guest joins the conversation
When the standalone signaling server is used and a participant joins a
conversation a "join" signaling message is received. The code assumed
that the event always contains a "user" attribute, but that attribute is
empty for guests. As the "user" attribute was used unconditionally this
caused an exception, and as the exception happened when handling a
websocket message it propagated until it caused a reconnection of the
websocket.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>