update version of dav4jvm to 2.1.2
Version 1.0 of dav4jvm had a dependency on okio 2.+
This led to the use of okio 2.10.0-SNAPSHOT, which caused "Unresolved reference: ByteString" so talk-android failed to build.
This commit sets version 2.1.2 for dav4jvm, which even has no direct dependency for okio any more (but for okhttp instead, which still uses okio 2.9.0).

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>