disable text fields for location messages (workaround)
i tried to suppress that zip codes are shown as clickable telephopne numbers. but
  android:autoLink="none"
and
  messageText!!.linksClickable = false
don't seem to work.
as a workaround i disabled the text field completely. so the links are still underlined but nut clickable. and it's also not possible now to mark text and get the context menu.

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>