stop mediaPlayer in onDestroy
a voice message now continues to play if app goes to background, but whenever the user returns he can pause or stop the playback (UI is updated).
Also, the playback survives a screen orientation change.

instead stopping the mediaPlayer in onDetach would stop playback when moving app to background but also stop the playback on a screen orientation change.

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>