Add missing .editorconfig
If an developer have a own '.editorconfig' in some location above the repostires
directory it will used in the gradle build and formatting checks can be fail.

To avoid this an get an more uniform code style between Nextclouds Android apps
the '.editorconfig' from the Nextcloud Android app is used.

Because of defining 'root=true' the developers own '.editorconfig' is not used
anymore.

Signed-off-by: Tim Krüger <t@timkrueger.me>