Remove media in text message detection
The method 'TextMatchers#getMessageTypeFromString' suggested MIME types
for linked files in text message. Also for links like this not working
example:

  - http://example.local/image.png
  - image.png

In those cases no image can be loaded and that results in empty previews
like shown in the issue #1167.

So for alignment reasons to spreed and Talk iOS app this media detection
is removed and only links will be shown.

Resolves: #1167

Signed-off-by: Tim Krüger <t@timkrueger.me>