Fix call flags in video calls
The call flags describe the streams provided by the client, so in a
video call both audio and video are provided, not just video.

Note that as publishing permissions are currently not implemented in the
Android app the provided flags do not take into account the available
permissions. Nevertheless, the server restricts the flags based on the
permissions, so the other participants would see the appropriate flags.
In the future it would be better to send the right flags directly from
the Android app.

Similarly, for now, it is just assumed that both audio and video tracks
will be provided by the Android app, but the flags should reflect the
actually available tracks (for example, if it was not possible to get a
video track for some reason the call flags should not include
"WITH_VIDEO").

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>